+++ atmosphere/tasks/kubernetes/v1.py
@@ -6,21 +6,6 @@ from atmosphere.tasks.kubernetes import base
 LOG = logger.get_logger()
 
 
-class ApplyNamespaceTask(base.ApplyKubernetesObjectTask):
-    def __init__(self, name: str):
-        super().__init__(kind=pykube.Namespace, namespace=None, name=name)
-
-    def generate_object(self) -> pykube.Namespace:
-        return pykube.Namespace(
-            self.api,
-            {
-                "apiVersion": self._obj_kind.version,
-                "kind": self._obj_kind.kind,
-                "metadata": {"name": self._obj_name},
-            },
-        )
-
-
 class ApplyConfigMapTask(base.ApplyKubernetesObjectTask):
     def __init__(self, namespace: str, name: str, data: str):
         self._data = data
@@ -29,7 +14,6 @@ class ApplyConfigMapTask(base.ApplyKubernetesObjectTask):
             kind=pykube.ConfigMap,
             namespace=namespace,
             name=name,
-            requires=set(["namespace"]),
         )
 
     def generate_object(self) -> pykube.ConfigMap:
@@ -56,7 +40,6 @@ class ApplyServiceTask(base.ApplyKubernetesObjectTask):
             kind=pykube.Service,
             namespace=namespace,
             name=name,
-            requires=set(["namespace"]),
         )
 
     def generate_object(self) -> pykube.Service:
@@ -92,7 +75,6 @@ class ApplyIngressTask(base.ApplyKubernetesObjectTask):
             kind=pykube.Ingress,
             namespace=namespace,
             name=name,
-            requires=set(["namespace"]),
         )
 
     def generate_object(self) -> pykube.Ingress:
@@ -120,7 +102,6 @@ class ApplySecretTask(base.ApplyKubernetesObjectTask):
             kind=pykube.Secret,
             namespace=namespace,
             name=name,
-            requires=set(["namespace"]),
         )
 
     def generate_object(self) -> pykube.Secret:
@@ -134,42 +115,5 @@ class ApplySecretTask(base.ApplyKubernetesObjectTask):
                     "namespace": self._obj_namespace,
                 },
                 "stringData": self._data,
-            },
-        )
-
-
-class ApplyIngressTask(base.ApplyKubernetesObjectTask):
-    def __init__(
-        self,
-        namespace: str,
-        name: str,
-        spec: dict,
-        annotations: dict = {},
-        labels: dict = {},
-    ):
-        self._annotations = annotations
-        self._labels = labels
-        self._spec = spec
-
-        super().__init__(
-            kind=pykube.Ingress,
-            namespace=namespace,
-            name=name,
-            requires=set(["namespace"]),
-        )
-
-    def generate_object(self) -> pykube.Ingress:
-        return pykube.Ingress(
-            self.api,
-            {
-                "apiVersion": self._obj_kind.version,
-                "kind": self._obj_kind.kind,
-                "metadata": {
-                    "name": self._obj_name,
-                    "namespace": self._obj_namespace,
-                    "annotations": self._annotations,
-                    "labels": self._labels,
-                },
-                "spec": self._spec,
             },
         )