+++ internal/pkg/image_repositories/build_workflow.go
@@ -32,10 +32,11 @@ var DIST_PACAKGES map[string]string = map[string]string{
 	"nova":          "ovmf qemu-efi-aarch64 lsscsi nvme-cli sysfsutils udev util-linux ndctl",
 }
 var PIP_PACKAGES map[string]string = map[string]string{
+	"cinder":        "purestorage",
 	"glance":        "glance_store[cinder]",
 	"horizon":       "git+https://github.com/openstack/designate-dashboard.git@stable/${{ matrix.release }} git+https://github.com/openstack/heat-dashboard.git@stable/${{ matrix.release }} git+https://github.com/openstack/ironic-ui.git@stable/${{ matrix.release }} git+https://github.com/openstack/magnum-ui.git@stable/${{ matrix.release }} git+https://github.com/openstack/neutron-vpnaas-dashboard.git@stable/${{ matrix.release }} git+https://github.com/openstack/octavia-dashboard.git@stable/${{ matrix.release }} git+https://github.com/openstack/senlin-dashboard.git@stable/${{ matrix.release }} git+https://github.com/openstack/monasca-ui.git@stable/${{ matrix.release }}",
 	"ironic":        "python-dracclient sushy",
-	"magnum":        "magnum-cluster-api==0.1.2",
+	"magnum":        "magnum-cluster-api==0.2.6",
 	"monasca-agent": "libvirt-python python-glanceclient python-neutronclient python-novaclient py3nvml",
 	"neutron":       "neutron-vpnaas",
 	"placement":     "httplib2",
@@ -77,6 +78,8 @@ func NewBuildWorkflow(project string) *GithubWorkflow {
 	}
 
 	buildArgs := []string{
+		"BUILDER_IMAGE=quay.io/vexxhost/openstack-builder-${{ matrix.from }}",
+		"RUNTIME_IMAGE=quay.io/vexxhost/openstack-runtime-${{ matrix.from }}",
 		"RELEASE=${{ matrix.release }}",
 		fmt.Sprintf("PROJECT=%s", project),
 		fmt.Sprintf("PROJECT_REPO=%s", gitRepo),
@@ -105,8 +108,23 @@ func NewBuildWorkflow(project string) *GithubWorkflow {
 			"image": {
 				RunsOn: "ubuntu-latest",
 				Strategy: GithubWorkflowStrategy{
-					Matrix: map[string][]string{
-						"release": {"wallaby", "xena", "yoga", "zed"},
+					Matrix: map[string]interface{}{
+						"from":    []string{"focal", "jammy"},
+						"release": []string{"wallaby", "xena", "yoga", "zed"},
+						"exclude": []map[string]string{
+							{
+								"from":    "focal",
+								"release": "zed",
+							},
+							{
+								"from":    "jammy",
+								"release": "wallaby",
+							},
+							{
+								"from":    "jammy",
+								"release": "xena",
+							},
+						},
 					},
 				},
 				Steps: []GithubWorkflowStep{
@@ -141,20 +159,20 @@ func NewBuildWorkflow(project string) *GithubWorkflow {
 						Uses: "docker/build-push-action@v3",
 						With: map[string]string{
 							"context":    ".",
-							"cache-from": "type=gha,scope=${{ matrix.release }}",
-							"cache-to":   "type=gha,mode=max,scope=${{ matrix.release }}",
+							"cache-from": "type=gha,scope=${{ matrix.from }}-${{ matrix.release }}",
+							"cache-to":   "type=gha,mode=max,scope=${{ matrix.from }}-${{ matrix.release }}",
 							"platforms":  platforms,
 							"push":       "${{ github.event_name == 'push' }}",
 							"build-args": strings.Join(buildArgs, "\n"),
-							"tags":       fmt.Sprintf("quay.io/vexxhost/%s:${{ env.PROJECT_REF }}", project),
+							"tags":       fmt.Sprintf("quay.io/vexxhost/%s:${{ env.PROJECT_REF }}-${{ matrix.from }}", project),
 						},
 					},
 					{
 						Name: "Promote image",
 						Uses: "akhilerm/tag-push-action@v2.0.0",
-						If:   "github.ref == 'refs/heads/main'",
+						If:   `github.event_name == 'push' && ((matrix.from == 'focal') || (matrix.from == 'jammy' && matrix.release != 'yoga'))`,
 						With: map[string]string{
-							"src": fmt.Sprintf("quay.io/vexxhost/%s:${{ env.PROJECT_REF }}", project),
+							"src": fmt.Sprintf("quay.io/vexxhost/%s:${{ env.PROJECT_REF }}-${{ matrix.from }}", project),
 							"dst": fmt.Sprintf("quay.io/vexxhost/%s:${{ matrix.release }}", project),
 						},
 					},